@@ -33,25 +33,26 @@ dependencies:
   injectable: ^2.5.0
   dartz: ^0.10.1
   # database written in pure Dart
-  hive: ^2.2.3
-  hive_flutter: ^1.1.0
+
   flutter_screenutil: ^5.9.0
-  freezed_annotation: ^2.4.4
+  freezed_annotation: ^3.1.0
   json_annotation: ^4.9.0
   flutter_launcher_icons: ^0.14.4
   firebase_messaging: ^15.2.10
   flutter_local_notifications: ^19.4.1
   retrofit: ^4.9.0
+  hive_ce: ^2.15.1
+  hive_ce_generator: ^1.10.0
+  hive_ce_flutter: ^2.3.3
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  injectable_generator: 2.6.2
+  injectable_generator: 2.9.1
   flutter_lints: ^5.0.0
   build_runner: ^2.5.4
-  hive_generator: ^2.0.0
   json_serializable: ^6.9.0
-  retrofit_generator: ^9.2.0
-  freezed: ^2.5.7
+  retrofit_generator: ^10.1.0
+  freezed: ^3.2.3
 flutter_launcher_icons:
   android: true
   ios: true
