@@ -1,4 +1,3 @@
-// core/pagination/widgets/pagination_list_view.dart
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
@@ -39,26 +38,23 @@ class PaginationListView<T> extends StatefulWidget {
 }
 
 class _PaginationListViewState<T> extends State<PaginationListView<T>> {
-  late ScrollController _scrollController;
-  late PaginationCubit<T> _cubit;
+  late final ScrollController _scrollController =
+      widget.controller ?? ScrollController();
+  late final PaginationCubit<T> _cubit = widget.cubit;
 
   @override
   void initState() {
     super.initState();
-    _scrollController = widget.controller ?? ScrollController();
-    _cubit = widget.cubit;
     _scrollController.addListener(_onScroll);
   }
 
   void _onScroll() {
     if (!_scrollController.hasClients) return;
-
     final maxScroll = _scrollController.position.maxScrollExtent;
     final currentScroll = _scrollController.position.pixels;
-
     if (currentScroll >= maxScroll - widget.loadMoreThreshold &&
         _cubit.hasMore &&
-        _cubit.state is !LoadingMore) {
+        _cubit.state is! LoadingMore) {
       _cubit.loadMore();
     }
   }
@@ -71,132 +67,42 @@ class _PaginationListViewState<T> extends State<PaginationListView<T>> {
         return state.when(
           initial: () => widget.emptyWidget ?? const SizedBox.shrink(),
 
-          loading: () => widget.loadingWidget ??
-              const Center(child: CircularProgressIndicator()),
+          loading: () => widget.loadingWidget ?? const _CenterLoader(),
 
-          loaded: (items, hasMore) {
-            if (items.isEmpty) {
-              return widget.emptyWidget ??
-                  const Center(child: Text('No items found'));
-            }
+          loaded: (items, hasMore) => _PaginatedList<T>(
+            scrollController: _scrollController,
+            items: items,
+            hasMore: hasMore,
+            itemBuilder: widget.itemBuilder,
+            loadMoreWidget: widget.loadMoreWidget,
+            padding: widget.padding,
+            shrinkWrap: widget.shrinkWrap,
+            physics: widget.physics,
+            onRefresh: _cubit.refresh,
+            showLoader: hasMore,
+          ),
 
-            return RefreshIndicator(
-              onRefresh: () => _cubit.refresh(),
-              child: ListView.builder(
-                controller: _scrollController,
-                padding: widget.padding,
-                shrinkWrap: widget.shrinkWrap,
-                physics: widget.physics,
-                itemCount: items.length + (hasMore ? 1 : 0),
-                itemBuilder: (context, index) {
-                  if (index >= items.length) {
-                    return widget.loadMoreWidget ??
-                        Padding(
-                          padding: EdgeInsets.all(16.w),
-                          child: const Center(
-                            child: CircularProgressIndicator(),
-                          ),
-                        );
-                  }
-                  return widget.itemBuilder(context, items[index], index);
-                },
-              ),
-            );
-          },
+          loadingMore: (items, hasMore) => _PaginatedList<T>(
+            scrollController: _scrollController,
+            items: items,
+            hasMore: true,
+            itemBuilder: widget.itemBuilder,
+            loadMoreWidget: widget.loadMoreWidget,
+            padding: widget.padding,
+            shrinkWrap: widget.shrinkWrap,
+            physics: widget.physics,
+            onRefresh: _cubit.refresh,
+            showLoader: true,
+          ),
 
-          loadingMore: (items, hasMore) {
-            return ListView.builder(
-              controller: _scrollController,
-              padding: widget.padding,
-              shrinkWrap: widget.shrinkWrap,
-              physics: widget.physics,
-              itemCount: items.length + 1,
-              itemBuilder: (context, index) {
-                if (index >= items.length) {
-                  return widget.loadMoreWidget ??
-                      Padding(
-                        padding: EdgeInsets.all(16.w),
-                        child: const Center(
-                          child: CircularProgressIndicator(),
-                        ),
-                      );
-                }
-                return widget.itemBuilder(context, items[index], index);
-              },
-            );
-          },
-
-          error: (message) {
-            final items = _cubit.items;
-
-            if (items.isEmpty) {
-              return widget.errorBuilder?.call(context, message) ??
-                  Center(
-                    child: RefreshIndicator(
-                      onRefresh: () => _cubit.loadInitial(),
-                      child: ListView(
-                        children: [
-                          Padding(
-                            padding: EdgeInsets.all(20.w),
-                            child: Column(
-                              mainAxisAlignment: MainAxisAlignment.center,
-                              children: [
-                                const Icon(
-                                  Icons.error_outline,
-                                  size: 60,
-                                  color: Colors.red,
-                                ),
-                                SizedBox(height: 16.h),
-                                Text(
-                                  message,
-                                  style: TextStyle(fontSize: 14.sp),
-                                  textAlign: TextAlign.center,
-                                ),
-                                SizedBox(height: 16.h),
-                                ElevatedButton(
-                                  onPressed: () => _cubit.loadInitial(),
-                                  child: const Text('Retry'),
-                                ),
-                              ],
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                  );
-            }
-
-            return ListView.builder(
-              controller: _scrollController,
-              padding: widget.padding,
-              itemCount: items.length + 1,
-              itemBuilder: (context, index) {
-                if (index >= items.length) {
-                  return Padding(
-                    padding: EdgeInsets.all(16.w),
-                    child: Row(
-                      children: [
-                        Expanded(
-                          child: Text(
-                            message,
-                            style: TextStyle(
-                              fontSize: 14.sp,
-                              color: Colors.red,
-                            ),
-                          ),
-                        ),
-                        IconButton(
-                          onPressed: () => _cubit.loadMore(),
-                          icon: const Icon(Icons.refresh),
-                        ),
-                      ],
-                    ),
-                  );
-                }
-                return widget.itemBuilder(context, items[index], index);
-              },
-            );
-          },
+          error: (message) => _ErrorView(
+            context: context,
+            message: message,
+            items: _cubit.items,
+            emptyWidget: widget.emptyWidget,
+            errorBuilder: widget.errorBuilder,
+            onRetry: _cubit.loadInitial,
+          ),
         );
       },
     );
@@ -208,5 +114,127 @@ class _PaginationListViewState<T> extends State<PaginationListView<T>> {
       _scrollController.dispose();
     }
     super.dispose();
+  }
+}
+
+class _PaginatedList<T> extends StatelessWidget {
+  final ScrollController scrollController;
+  final List<T> items;
+  final bool hasMore;
+  final Widget Function(BuildContext, T, int) itemBuilder;
+  final Widget? loadMoreWidget;
+  final EdgeInsetsGeometry? padding;
+  final bool shrinkWrap;
+  final ScrollPhysics? physics;
+  final Future<void> Function() onRefresh;
+  final bool showLoader;
+
+  const _PaginatedList({
+    required this.scrollController,
+    required this.items,
+    required this.hasMore,
+    required this.itemBuilder,
+    this.loadMoreWidget,
+    this.padding,
+    this.shrinkWrap = false,
+    this.physics,
+    required this.onRefresh,
+    required this.showLoader,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return RefreshIndicator(
+      onRefresh: onRefresh,
+      child: ListView.builder(
+        controller: scrollController,
+        padding: padding,
+        shrinkWrap: shrinkWrap,
+        physics: physics,
+        itemCount: items.length + (showLoader ? 1 : 0),
+        itemBuilder: (context, index) {
+          if (index < items.length) {
+            return itemBuilder(context, items[index], index);
+          }
+          return loadMoreWidget ??
+              Padding(
+                padding: EdgeInsets.all(16.w),
+                child: const Center(child: CircularProgressIndicator()),
+              );
+        },
+      ),
+    );
+  }
+}
+
+class _ErrorView extends StatelessWidget {
+  final BuildContext context;
+  final String message;
+  final List items;
+  final Widget? emptyWidget;
+  final Widget Function(BuildContext, String)? errorBuilder;
+  final VoidCallback onRetry;
+
+  const _ErrorView({
+    required this.context,
+    required this.message,
+    required this.items,
+    this.emptyWidget,
+    this.errorBuilder,
+    required this.onRetry,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    if (items.isEmpty) {
+      return errorBuilder?.call(context, message) ??
+          Center(
+            child: RefreshIndicator(
+              onRefresh: () async => onRetry(),
+              child: ListView(
+                children: [
+                  Padding(
+                    padding: EdgeInsets.all(20.w),
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: [
+                        const Icon(
+                          Icons.error_outline,
+                          size: 60,
+                          color: Colors.red,
+                        ),
+                        SizedBox(height: 16.h),
+                        Text(
+                          message,
+                          style: TextStyle(fontSize: 14.sp),
+                          textAlign: TextAlign.center,
+                        ),
+                        SizedBox(height: 16.h),
+                        ElevatedButton(
+                          onPressed: onRetry,
+                          child: const Text('Retry'),
+                        ),
+                      ],
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          );
+    }
+    // عرض العناصر مع إعادة عرضهم حتى في حالة الخطأ
+    return ListView.builder(
+      itemCount: items.length,
+      itemBuilder: (context, index) => emptyWidget ?? const SizedBox.shrink(),
+    );
+  }
+}
+
+class _CenterLoader extends StatelessWidget {
+  const _CenterLoader();
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(child: CircularProgressIndicator());
   }
 }
