@@ -2,6 +2,7 @@
 import 'package:bloc/bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:injectable/injectable.dart';
+import 'package:wasil_task/core/network/api_result.dart';
 import '../../../domain/entites/product_details_entity.dart';
 import '../../../domain/use_cases/get_product_details_usecase.dart';
 
@@ -13,7 +14,7 @@ class ProductDetailsCubit extends Cubit<ProductDetailsState> {
   final GetProductDetailsUseCase productDetailsUseCase;
 
   ProductDetailsCubit(this.productDetailsUseCase)
-      : super(const ProductDetailsState.initial());
+    : super(const ProductDetailsState.initial());
 
   Future<void> getProductDetails(int id) async {
     emit(const ProductDetailsState.loading());
@@ -21,9 +22,7 @@ class ProductDetailsCubit extends Cubit<ProductDetailsState> {
     final result = await productDetailsUseCase(id);
 
     result.when(
-      success: (product) => emit(
-        ProductDetailsState.success(product: product),
-      ),
+      success: (product) => emit(ProductDetailsState.success(product: product)),
       failure: (error) => emit(
         ProductDetailsState.failure(message: error.message ?? 'Unknown error'),
       ),
